refactor(synapsys): extract domain helpers to shared module
Move getDomains, setsIntersect, firstShared, forEachPair to new lib/lint/domain-utils.js to eliminate duplicate-blocks violations in synapsys-lint.js and pair-scoring.js. Both pretool-scoring.js and pair-scoring.js now use shared forEachPair for (i<j) pair iteration.
